updated visualization.  add items on addHeroData
visualization can show damage or battletag.  need to fix so both shows
up with different colors.  when a hero is inserted, so are the items.
all heroes are added when loading the player page, as a result all items
are added.  need to handle when there are too many requests.