getItemIDsFromHero takes in less parameters
d3functions.js:
--error handling adding if extraItemData was undefined
--removed addLegend().  its functionality is handled outside svg.
--renamed stats, and items to statsd3, itemsd3

gemMethods.js:
--renamed requestedRankHigher --> requestedGemRankHigher

heroMethods.js:
--added/removed comments
--findItemsInCollection takes in a hero object that contains heroID,
class, and items instead of heroID, and allItems.  less parameters and
allows class of hero to be accessed.
--compareItems()-changed ring & amulet logic when (equipped gem was not
boon, and request was not a higher rank)  ADD check to see if same rank.

playerMethods.js:
--removed used lines

renderMethods.js:
--changed heroMethods.getItemIDsFromHero to take in a hero object

classLeaderboard.ejs:
--set dps span to have class currentGraph on document ready

hero.ejs:
--lastupdated was implemented later and therefore caused pages to not
render because not all hero's had a lastupdated.  fixed it to show
"update now!" if lastupdated was undefined